cleaning unused class
